+++ ops/testnet/prod/backend/config.tf
@@ -10,7 +10,8 @@ locals {
     { name = "PGRST_DB_URI", value = "postgres://${var.postgres_user}:${var.postgres_password}@${module.poller_db.db_instance_endpoint}/connext" },
     { name = "PGRST_DB_SCHEMA", value = "public" },
     { name = "PGRST_DB_ANON_ROLE", value = "query" },
-    { name = "ENVIRONMENT", value = var.environment }
+    { name = "ENVIRONMENT", value = var.environment },
+    { name = "STAGE", value = var.stage }
   ]
 }
 
@@ -21,6 +22,6 @@ locals {
       "1111" = {}
       "2221" = {}
     }
-    environment = "production"
+    environment = var.stage
   })
 }